pyproject.toml
[project]
name = "ai-company"
version = "0.1.0"
description = "AI-driven R&D organization management system"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
